requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
requests==2.31.0
stripe==7.4.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
django-allauth==0.57.0
Pillow==10.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.22
django-otp==1.2.2
django-two-factor-auth==1.15.5
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.2
boto3==1.33.11
celery[redis]==5.3.4
flower==2.0.1
django-debug-toolbar==4.2.0
python-decouple==3.8
openpyxl==3.1.2
pandas==2.1.3
python-magic-bin==0.4.14  # Windows-friendly version
django-celery-beat==2.5.0
django-celery-results==2.5.1